Use views instead of native queries where possible (#33)
This PR changes a lot about how glvd-api works.

The biggest change is that we now rely on VIEWs in the db instead of having native queries in our repositories where possible. This simplifies the code a lot and should also make it easier to implement sorting https://github.com/gardenlinux/glvd/issues/95 and filtering https://github.com/gardenlinux/glvd/issues/94 in the api.