Use kernel.org vulnerabilities repo for kernel CVEs (#95)
The linux kernel vulnerabilities we get from the debian security tracker are not useful for Garden Linux because we maintain our own build of the LTS kernel.
This PR makes use of kernel CVE information we obtain from kernel.org for that specific LTS line of the kernel.

Fixes https://github.com/gardenlinux/glvd/issues/122

This PR requires https://github.com/gardenlinux/glvd-data-ingestion/pull/31 to be merged and the schema to be updated to that version.