Don't include linux kernel CVEs when list of packages contains 'linux' in a package name (#120)
Due to the way we handle linux kernel cves we need special handling for them. In the linked issue, it was noticed that our way of identifying the linux kernel in the list of provided packages was too naive.

This PR adds test cases to reproduce the issue, and expands the condition to one that should work in all cases.

Fixes https://github.com/gardenlinux/glvd/issues/167