Ignore CVEs with status 'Rejected' in list of vulnerabilities (#143)
CVEs with status 'Rejected' are typically not those people care about. GLVD will mark a CVE as vulnerable if there is no fixed version, which might be the case for rejected CVEs. This PR add the `vulnStatus` field to the api, and filters 'rejected' CVEs directly in the API. CVE Details for rejected CVEs can still be viewed.

Fixes https://github.com/gardenlinux/glvd/issues/182